+++ extension/package.json
@@ -25,7 +25,6 @@
   },
   "devDependencies": {
     "@gnosis.pm/zodiac": "^2.0.1",
-    "@cowprotocol/cow-sdk": "^2.0.0-alpha.4",
     "@safe-global/safe-core-sdk": "^3.2.0",
     "@safe-global/safe-core-sdk-types": "^1.7.0",
     "@safe-global/safe-ethers-lib": "^1.7.0",
