+++ extension/src/transactionTranslations/cowswapSetPreSignature.ts
@@ -1,17 +1,14 @@
+import { Web3Provider } from '@ethersproject/providers'
 import { KnownContracts } from '@gnosis.pm/zodiac'
-import { Interface } from 'ethers/lib/utils'
-import { TransactionTranslation } from './types'
-import { getDefaultProvider } from '@ethersproject/providers'
 import { ethers } from 'ethers'
-// import { CowSwap } from '@cowprotocol/cow-sdk'
+
+import { TransactionTranslation } from './types'
 
 export const COWSWAP_SUPPORTED_NETWORK: Record<number, string> = {
   1: 'mainnet',
   5: 'goerli',
   100: 'xdai',
 }
-
-const API_URL = `https://api.cow.fi/`
 
 const abi = [
   {
@@ -270,7 +267,7 @@ const abi = [
     outputs: [{ type: 'address' }],
   },
 ]
-const provider = getDefaultProvider()
+const provider = window.ethereum && new Web3Provider(window.ethereum)
 const cowSwapFunctionSignature = 'setPreSignature(bytes,bool)'
 const cowSwapEncodeFunctionSignature = ethers.utils
   .id(cowSwapFunctionSignature)
@@ -283,41 +280,46 @@ export default {
 
   translate: async (transaction) => {
     let preSignature = undefined
-    const { chainId } = await provider.getNetwork()
-    const COW_SWAP_URL = `${API_URL}/${COWSWAP_SUPPORTED_NETWORK[chainId]}/api/v1/orders`
+
     if (!transaction.data) {
       return undefined
     }
 
-    const iface = new ethers.utils.Interface(abi)
-    const transactionDecoded = iface.parseTransaction({
-      data: transaction.data,
-      value: transaction.value,
-    })
-    if (transactionDecoded.sighash === cowSwapEncodeFunctionSignature) {
-      const cowSwapInputValues = iface.decodeFunctionData(
-        cowSwapFunctionSignature,
-        transaction.data
-      )
-      if (cowSwapInputValues.length) {
-        const orderUid = cowSwapInputValues[0]
-        if (orderUid) {
-          try {
-            await fetch(`${COW_SWAP_URL}/${orderUid}`) //api for the get request
-              .then((response) => response.json())
-              .then((data) => {
-                if (data.status === 'presignaturePending') {
-                  preSignature = [
-                    {
-                      ...transaction,
-                      value: transaction.value,
-                      data: transaction.data,
-                    },
-                  ]
-                }
-              })
-          } catch (e) {
-            throw new Error(`Error fetching the ${orderUid} from cowswap API`)
+    if (provider) {
+      const chainId = provider.network.chainId
+      const iface = new ethers.utils.Interface(abi)
+      const transactionDecoded = iface.parseTransaction({
+        data: transaction.data,
+        value: transaction.value,
+      })
+
+      if (transactionDecoded.sighash === cowSwapEncodeFunctionSignature) {
+        const cowSwapInputValues = iface.decodeFunctionData(
+          cowSwapFunctionSignature,
+          transaction.data
+        )
+        if (cowSwapInputValues.length) {
+          const orderUid = cowSwapInputValues[0]
+          if (orderUid && chainId) {
+            try {
+              const API_URL = `https://api.cow.fi`
+              const COW_SWAP_URL = `${API_URL}/${COWSWAP_SUPPORTED_NETWORK[chainId]}/api/v1/orders`
+              await fetch(`${COW_SWAP_URL}/${orderUid}`) //api for the get request
+                .then((response) => response.json())
+                .then((data) => {
+                  if (data.status === 'presignaturePending') {
+                    preSignature = [
+                      {
+                        ...transaction,
+                        value: transaction.value,
+                        data: transaction.data,
+                      },
+                    ]
+                  }
+                })
+            } catch (e) {
+              throw new Error(`Error fetching the ${orderUid} from cowswap API`)
+            }
           }
         }
       }
